linkedList

+++ try/split.cpp
@@ -5,7 +5,7 @@ void split(string str, char del){
 	// declaring temp string to store the curr "word" upto del
 	string temp = "";
 
-	for(int i=0; i<(int)str.size(); i++){
+	for(int i=0; i<str.size(); i++){
 		// If cur char is not del, then append it to the cur "word", otherwise
 		// you have completed the word, print it, and start a new word.
 		if(str[i] != del){
@@ -16,7 +16,6 @@ void split(string str, char del){
 			temp = "";
 		}
 	}
-
 	cout << temp;
 }
